+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ Change Log
 * Added check for ES_HOST_LOCAL environment variable in ff_utils.get_es_metadata;
   for running Foursight checks locally (with local ssh tunnel to ES proxy);
   came up in foursight/checks/audit_checks (2024-04-23).
+* Added extract_file_from_zip to zip_utils.
+* Added http_utils with download function.
 
 
 8.8.4